+++ internal/services/serviceprincipals/service_principal_delegated_permission_grant_resource.go
@@ -41,11 +41,11 @@ func servicePrincipalDelegatedPermissionGrantResource() *pluginsdk.Resource {
 
 		Importer: pluginsdk.ImporterValidatingResourceId(func(id string) error {
 			if _, errs := stable.ValidateOAuth2PermissionGrantID(id, "id"); len(errs) > 0 {
-				out := ""
+				var out strings.Builder
 				for _, err := range errs {
-					out += err.Error()
+					out.WriteString(err.Error())
 				}
-				return errors.New(out)
+				return errors.New(out.String())
 			}
 			return nil
 		}),