+++ internal/services/userflows/user_flow_attribute_resource_test.go
@@ -35,10 +35,8 @@ func TestAccUserFlowAttributeResource_serialised(t *testing.T) {
 	}
 
 	for group, m := range testCases {
-		m := m
 		t.Run(group, func(t *testing.T) {
 			for name, tc := range m {
-				tc := tc
 				t.Run(name, func(t *testing.T) {
 					tc(t)
 				})